good path, good plot, still bad edges

+++ guidance/find_path/debug_plot.py
@@ -1,6 +1,7 @@
 import cv2 as cv
 import numpy as np
 import matplotlib.pyplot as plt
+import matplotlib.image as image
 import math
 import imageio.v2 as imageio
 from random import randint
@@ -10,6 +11,7 @@ DEBUG_PRINT = False
 PLOT = True
 DRAW_RADIUS = 10
 TEXT_SIZE = 1.2
+POINTS_SIZE = 2
 
 ########################### DEBUG ###########################
 def outOfBorders(src, point):
@@ -146,47 +148,64 @@ def plotClusters(clusters, shape, title=None, draw_points = False, pointRadius =
         title = str(len(clusters)) + ' Clusters'
     myPlot(fig, title)
 
-def plotPoints(points, sk, gifname = 'trajectory', remove_files = True):
+def plotPoints(map, points, start_id, end_id, img_name, save_name = 'trajectory', gif = False, remove_files = True, n_points = 100):
     """plot points on image and save trajectory as a gif."""
 
-    filenames = ['sk.png']
-    plt.imshow(sk, cmap='gray_r')
-    plt.savefig(filenames[0])
+    filenames = ['original.png']
+    if gif:
+        n_points = 30
+    # downsample points for speed (if gif, even more)
+    points = [points[i] for i in range(0, len(points), int(len(points) / n_points))]
 
-    for i,(y,x) in enumerate(points):
+    # plot image represented in file img_name as background in plt object
+    img = cv.imread(img_name)
+    plt.imshow(img)
+
+    color = 'red'
+    for i,(x,y) in enumerate(points):
         filename = f'{i}.png'
         filenames.append(filename)
-        a,b,c = num_to_rgb(i, len(points))
-        plt.scatter( x, y, c =  np.array([[a, b, c]])/255.0)
-        plt.savefig(filename)
-        # plt.pause(0.01)
-        i+=1
+        # plot points with size 1
+        plt.scatter( x, y, c = color, s=POINTS_SIZE)
+        if gif:
+            plt.savefig(filename)
+
+    for end in map.unique_ends:
+        x,y = end.location
+        plt.text(x, y, end.id, color='blue')
+
+    plt.title('Trajectory from ' + str(start_id) + ' to ' + str(end_id))
+    plt.savefig(save_name + '_path' + '.png')
 
     if PLOT:
-        plt.title('Points')
         plt.show()
 
     # build gif
-    with imageio.get_writer(gifname + '.gif', mode='I') as writer:
-        for filename in filenames:
-            image = imageio.imread(filename)
-            writer.append_data(image)         
-    # remove files
-    if remove_files:
-        for filename in set(filenames):
-            os.remove(filename)
+    if gif:
+        with imageio.get_writer(save_name + '.gif', mode='I') as writer:
+            for filename in filenames:
+                image = imageio.imread(filename)
+                writer.append_data(image)         
+        # remove files
+        if remove_files:
+            for filename in set(filenames):
+                os.remove(filename)
     
     plt.close()
 
-def myPlot(src, title, cmap='gray_r'):
+def myPlot(src, title, cmap='gray_r', save_name = None):
     """plot image"""
 
-    if not PLOT:
-        return
-    if src is not None:
+    if src is not None and PLOT:
         plt.imshow(src, cmap=cmap)
         plt.title(title)
         plt.show()
+        if save_name is not None:
+            plt.imsave(save_name+'.png', src, cmap=cmap)
+    elif src is not None and save_name is not None:
+        plt.imshow(src, cmap=cmap)
+        plt.title(title)
+        plt.imsave(save_name+'.png', src, cmap=cmap)
 
 def plotRelevantPoints(sk, starting_points, bif_points):
     """plot relevant points in skeleton"""
@@ -198,23 +217,30 @@ def plotRelevantPoints(sk, starting_points, bif_points):
     fig = drawPoints(fig, bif_points, color=(0,255,0), radius=DRAW_RADIUS)
     myPlot(fig, title)
 
-def plotMap(map, title = None):
+def plotMap(map, title = None, save_name = None):
     """plot map"""
     if not PLOT:
         return
 
-    edges = map.edges.copy()
-    shape = map.plot_shape    
+    all_edges = map.edges.copy()
+    all_edges.extend(map.bad_edges.copy())
+
     clusters = []
-    # switch x and y in clusters
-    for edge in edges:
+    shape = map.plot_shape
+
+    # switch x and y in clusters and ends locations
+    for edge in all_edges:
         edge.cluster = [(y,x) for (x,y) in edge.cluster]
         clusters.append(edge.cluster)
-    
-    # switch x and y in ends
-    for edge in edges:
         for end in edge.ends:
             end.location = (end.location[1], end.location[0])
+
+    # separate good and bad eges, good and bad clusters
+    bad_edges = all_edges[len(map.edges):]
+    edges = all_edges[0:len(map.edges)]
+    bad_clusters = clusters[len(map.edges):]
+    clusters = clusters[0:len(map.edges)]
+
 
     # create figure where cluster will be drawn
     fig = cv.cvtColor(np.full(shape, 255, dtype=np.uint8), cv.COLOR_GRAY2BGR)
@@ -223,9 +249,16 @@ def plotMap(map, title = None):
         # draw lines between consecutive cluster points
         for i in range(len(cluster)-1):
             cv.line(fig, cluster[i], cluster[i+1], num_to_rgb(color,len(clusters)), 3, cv.LINE_AA)
+    
+    # draw bad clusters in gray
+    color = (128,128,128)
+    for cluster in bad_clusters:
+        # draw lines between consecutive cluster points
+        for i in range(len(cluster)-1):
+            cv.line(fig, cluster[i], cluster[i+1], color, 3, cv.LINE_AA)
 
     # draw ends ids and edge distances
-    for edge in edges:
+    for edge in all_edges:
         for end in edge.ends:
             # print ends ids of edge in their location
             cv.putText(fig, end.id, end.location, cv.FONT_HERSHEY_SIMPLEX, TEXT_SIZE, (0,0,0), 1, cv.LINE_AA)
@@ -234,6 +267,6 @@ def plotMap(map, title = None):
         cv.putText(fig, str(round(edge.distance)), middle_point, cv.FONT_HERSHEY_SIMPLEX, TEXT_SIZE/2, (0,0,0), 1, cv.LINE_AA)
      
     if title is None:
-        title = str(len(edges)) + ' Edges'
+        title = str(len(edges)) + ' Edges, ' + str(len(bad_edges)) + ' Bad Edges'
 
-    myPlot(fig, title)
+    myPlot(fig, title, save_name = save_name)
